requirements.txt
streamlit>=1.36
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0
